NAN: Change the return value of nan_de_rx_sdf()
Change the return value of nan_de_rx_sdf() and relevant function
internal to the NAN DE to indicate if the information contained
in the SDF was valuable, e.g., there was a 'match' or 'result'
event.

This is needed in later patches that add logic to store information
about a peer, i.e., avoid storing peers that are not relevant.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>